Trim mdlFile 003204a0 callback width

+++ src/Battle/btlFormation.c
@@ -622,7 +622,7 @@ extern u32 func_002bc830(u8 *param_1);
 extern void func_002bc890(u8 *param_1,u16 param_2);
 extern u32 func_002bc910(u32 *param_1);
 extern void func_002bc950(u32 param_1,u32 param_2,u32 param_3);
-extern void func_002bc9c0(float param_1,float param_2,float param_3,float param_4,float param_5, float param_6,u32 param_7,long param_8);
+extern void func_002bc9c0(float param_1,float param_2,float param_3,float param_4,float param_5, float param_6,u32 param_7,int param_8);
 extern void func_002bccd0(u32 param_1, u32 param_2);
 extern void func_002bcde0(u64 param_1,u32 *param_2);
 extern void func_002bce10(float param_1,float param_2,u8* param_3,int param_4,float *param_5);
@@ -3370,7 +3370,7 @@ void func_002bc950(u32 param_1,u32 param_2,u32 param_3)
 // FUN_002bc9c0 NONMATCHING
 
 void func_002bc9c0(float param_1,float param_2,float param_3,float param_4,float param_5,
-                 float param_6,u32 param_7,long param_8)
+                 float param_6,u32 param_7,int param_8)
 
 {
   int iVar1 = 0;
